Add files via upload
used solidity to design and create a smart contract for a crowdfunding platform, which includes an organization raising money from the public for a specific objective, such as a project, charity, or business. The manager ,sets a target, deadline, and minimum contribution for the crowd. The crowd transfer funds to the smart contract, and the manager will need crowd approval to withdraw funds from the smart contract.